Move color_utils to _deprecated
